resize2fs: mark that the error return is deliberately ignored
When moving the inode table, if writing the (partially overlapping)
inode table fails, we need to write it back in its original location
before bailing out.  If that write unding the initial write fails,
there's nothing we can do, so we ignore it.  Mark this to avoid a
false positive from Coverity.

Fixes-Coverity-bug: 1432422
Signed-off-by: Theodore Ts'o <tytso@mit.edu>