Add a callback for overriding ext2fs_process_dir_block().
We have an application which creates a filesystem, creates many files in many directories on the filesystem, and performs no other operations on the filesystem. Creating a file in a directory is a quadratic-time operation using e2fsprogs becaus ext2fs_link() must do a linear scan of the directory to find the extent to put an inode in. For us this will always be the last extent.

The dominant operation in the linear directory scan is ext2fs_process_dir_block() because it performs I/O. The callback will let us keep track of the block number of the last extent in the directory and only call ext2fs_process_dir_block() on the last extent, except if the directory was just expanded using ext2fs_expand_dir(), in which case the full scan is necessary, but rare. This improves the performance of adding 100,000 files to a directory by 8x for us.