Improve wording on inaccurate warning message
When mounting a filesystem with fuse2fs, a warning message indicates
that writing to the journal is not supported. This can be interpreted as
the filesystem lacking a journal.

Clarify this message by being explicit in that fuse2fs does not support
journaling.

DRAFT: this patch is a draft; the POT file need to be regenerated, but
I'm not sure how to do so.

See: https://github.com/tytso/e2fsprogs/issues/220