fuse2fs: reopen filesystem read-write for read-only journal recovery
This changes the strategy added in c7f2688540d95e7f2cbcd178f8ff62ebe079faf7
for recovery of journals when read-only is requested.  That strategy always
opened the filesystem file read-write first, in case there was a journal to
recover.  A big problem with that strategy was that the user might not have
write access to the file.  The new strategy with read-only mounts is to
open the filesystem read-only first, then if there is a journal that needs
recovery, attempt to open it read-write.  If that works and the journal is
recovered, reopen it again read-only.

Signed-off-by: Dave Dykstra <2129743+DrDaveD@users.noreply.github.com>