fuse2fs: updates for message reporting journal is not supported
This makes two changes to the message that is shown saying that fuse2fs
does not support the journal.  First is that it reverts the check to
what it was before 3875380 to look at the ro option not being set
instead of checking the RW flag.  That's because I don't think this
message needs to be shown when the ro option is set even when it was
opened RW; there should be nothing to corrupt when it is ro.

Second, it changes the message to say that writing is not supported
rather than using the journal is not supported.  The current message is
confusing because in fact the journal is used for recovery when needed
and possible.

Signed-off-by: Dave Dykstra <2129743+DrDaveD@users.noreply.github.com>