fuse2fs: open read-only when image is not writable
This opens the image read-only when the image is not writable. If it is then found that a journal recovery is needed, an error is returned then.

The ret value is set to 2 after the option checks so that if there's an error resulting in "goto out" it won't print an error about unrecognized options.

Signed-off-by: Dave Dykstra <2129743+DrDaveD@users.noreply.github.com>